Lapic Timer Driver
Implement a timer driver using Lapic tsc deadline mode.

This driver enables the tsc deadline mode of Lapic if the platform supports
tsc deadline mode. Writing non-zero time interval in 100ns units arms the
timer. It writes a deadline to MSR_IA32_TSC_DEADLINE and the lapic gernerates
an interrupt when the tsc of cpu is greater or exceeds the deadline.

Signed-off-by: Yang, Yu-chu <yu-chu.yang@intel.com>